requirements.txt
python-telegram-bot>=20,<22
pillow
